ml-api/requirements.txt
# ml-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.0.1
torchvision==0.15.2
pillow==10.0.1
numpy==1.24.3


